Abstract r2d builder
